End desktop styles
