+++ src/pages/Training.jsx
@@ -1,13 +1,16 @@
 import ChartLine from 'components/ChartLine/ChartLine';
 import MyPurposeToRead from 'components/MyPurposeToRead/MyPurposeToRead';
 import { MyTraining } from 'components/MyTraining/MyTraining';
+import { Container } from '../App.styled';
 
 const Training = () => {
   return (
     <>
-      <MyTraining />
-      <MyPurposeToRead />
-      <ChartLine />
+      <Container>
+        <MyTraining />
+        <MyPurposeToRead />
+        <ChartLine />
+      </Container>
     </>
   );
 };
